+++ changelog.pl
@@ -1,5 +1,5 @@
 #
-# Authentic Theme 11.01 (https://github.com/qooob/authentic-theme)
+# Authentic Theme 11.10 (https://github.com/qooob/authentic-theme)
 # Copyright 2015 Ilia Rostovtsev <programming@rostovtsev.ru>
 # Licensed under MIT (https://github.com/qooob/authentic-theme/blob/master/LICENSE)
 #
@@ -13,27 +13,18 @@ our $__changelog
                 <h4 class="modal-title" id="update_notice_label">' . $text{'theme_update_notice'} . '</h4>
               </div>
               <div class="modal-body">
-                <h4>Version 11.00-11.01 (March 29, 2015)</h4>
+                <h4>Version 11.10 (April 13, 2015)</h4>
                 <ul>
-                    <li>Fixed unclickable left menu links, when custom logo is used <a href="https://github.com/qooob/authentic-theme/issues/143" target="_blank">#143</a></li>
-                    <li>Fixed DHCP-server module showing no hostnames <a href="https://github.com/qooob/authentic-theme/issues/145" target="_blank">#145</a></li>
-                    <li>Fixed multiple small bugs<hr></li>
-
-                    <li>Added window automatic scrolling, upon page is populated from server-side.Test it and see it in action in such modules as <em>Fetchmail Mail Retrieval</em>, <em>Software Package Updates</em> and others. Intended mouse-scroll done by user, during auto-scrolling, will make it stop. Loader will be automatically hidden upon this features is triggered. This feature can be disabled using settings</li>
-                    <li>Added in settings: <code>settings_security_notify_on_pre_login_request</code>, <code>settings_security_notify_on_login_request</code>, <code>settings_security_notify_on_login_success</code>, <code>settings_window_autoscroll</code>, <code>settings_sysinfo_theme_updates</code>, <code>settings_sysinfo_csf_updates</code>, <code>settings_leftmenu_button_language</code>, <code>settings_leftmenu_singlelink_icons</code>, <code>settings_leftmenu_vm_installscripts</code>, <code>settings_leftmenu_vm_webpages</code> and <code>settings_leftmenu_vm_backup_amazon</code>. Read the <a href="https://github.com/qooob/authentic-theme#how-do-i-use-theme-settings" target="_blank">manual</a> for using it</li>
-                    <li>Added automatically stretching inputs for long text <a href="https://github.com/qooob/authentic-theme/issues/121" target="_blank">#121</a></li>
-                    <li>Added auto-detection/redirection to <em>Mailbox</em> module upon logging in to <em>Usermin</em></li>
-                    <li>Added <em>Dutch</em> translation by <a href="https://github.com/Rvanlaak" target="_blank">Richard van Laak</a></li>
-                    <li>Added <em>Russian</em> translation</li>
-                    <li>Fixed appearing loader on changing <code>select</code>, even though it was disabled in settings <a href="https://github.com/qooob/authentic-theme/issues/127" target="_blank">#127</a></li>
-                    <li>Fixed text not being <em>wrapped</em> in <code>pre</code> tag and now scrollable horizontally instead <a href="https://github.com/qooob/authentic-theme/issues/123" target="_blank">#123</a></li>
-                    <li>Fixed left menu output issue in uncategorised mode <a href="https://github.com/qooob/authentic-theme/issues/128" target="_blank">#128</a></li>
-                    <li>Fixed encoding issue <a href="https://github.com/qooob/authentic-theme/issues/129" target="_blank">#129</a></li>
-                    <li>Fixed stuck loader in <em>Virtualmin</em>, while using third party modules</li>
-                    <li>Fixed left menu being shown irrespectively to the right frame <a href="https://github.com/qooob/authentic-theme/issues/131" target="_blank">#131</a></li>
-                    <li>Fixed external custom links not being opened in new window in <em>Virtualmin</em> module <a href="https://github.com/qooob/authentic-theme/issues/130" target="_blank">#130</a></li>
-                    <li>Improved UI of the right frame - dozens of improvements on fonts, tables, buttons, scrollbars and etc</li>
-                    <li>Remove all dependencies <a href="https://github.com/qooob/authentic-theme/issues/125" target="_blank">#125</a></li>
+                  <li>Added ability to <code>open new tab</code> of <em>Webmin/Usermin/Virtualmin/Cloudmin/Webmail</em> by clicking on the switch (top left menu\'s selector) by right mouse button.<br><strong>Important:</strong> You must once add <em>popup window exception</em> in your browser to make this feature work</li>
+                  <li>Improved greatly the usage of <code>autocomplete</code> for basic <em>Webmin</em> Modules. You can test it by going to <em>Running Processes</em> or <em>Perl Modules</em> and start typing in search field. Same done for most inbuilt, crucial modules</li>
+                  <li>Removed a hack for <em>Historic System Statistics</em> module, as it was fixed by the author in version 2.11 <a href="https://github.com/qooob/authentic-theme/issues/1" target="_blank">#1</a></li>
+                  <li>Fixed <code>login screen</code> missing for the link to <em>Virtualmin Password Recovery</em> module, in case it\'s installed and set in <em>Webmin Configuration->Anonymous Module Access</em>. In that case, the reset button will act as expected an be the color of yellow <a href="https://github.com/qooob/authentic-theme/issues/152" target="_blank">#152</a></li>
+                  <li>Fixed <em>Webmin Servers Index</em> module issue <a href="https://github.com/qooob/authentic-theme/issues/147" target="_blank">#147</a></li>
+                  <li>Fixed databases listing in editing table cells mode <a href="https://github.com/qooob/authentic-theme/issues/150" target="_blank">#150</a></li>
+                  <li>Fixed increased the height of <code>::-webkit-scrollbar</code> for easy access</li>
+                  <li>Fixed <code>.cgi</code> file attributes and set to be executable by default <a href="https://github.com/qooob/authentic-theme/issues/154" target="_blank">#154</a></li>
+                  <li>Fixed missing images in <em>Translator Module</em> <a href="https://github.com/qooob/authentic-theme/issues/155" target="_blank">#155</a></li>
+                  <li>Fixed multiple small bugs</li>
                 </ul>
                 <h4 style="margin-top:20px">' . $text{'theme_development_support'} . '</h4>
                 Thank you for using <a target="_blank" href="https://github.com/qooob/authentic-theme"><kbd style="background:#5cb85c">' . $text{'theme_name'} . '<kbd></a>. Overall development of this theme has already passed the stage of <kbd>400</kbd> hours.